feat(mail,kinds,hub,component,substrate): widen MailboxId u32 to u64 (adr-0029 phase 1) (#129)
## Summary
- Widens `MailboxId` and every wire/FFI field that carries one from
`u32` to `u64`, ahead of the ADR-0029 switch to hashed ids.
- Pure plumbing change: registry still allocates sequential ids; no
behavior change.
- `send_mail_p32`'s recipient param becomes `i64` at the wasm FFI
(scalar, not pointer — the `_p32` suffix still applies only to
pointer-typed imports per ADR-0024). `resolve_mailbox_p32` returns
`i64`. `MAILBOX_NOT_FOUND` widens to `u64::MAX`.
- Wire kinds widened: `LoadResult.mailbox_id`,
`DropComponent.mailbox_id`, `ReplaceComponent.mailbox_id`,
`SubscribeInput.mailbox`, `UnsubscribeInput.mailbox`.
- WAT fixtures (`end_to_end.rs`, `input_subscriptions.rs`,
`control.rs::WAT_FORWARDS_TO_SINK`) track the recipient-arg width bump
with `i64.const` / `i64.load8_u`.

## Test plan
- [x] `cargo test --workspace` — 93 substrate unit tests + 6 main bin +
1 end_to_end + 6 hub_client + 5 input_subscriptions all pass
- [x] `cargo clippy --all-targets -- -D warnings` clean
- [x] `cargo fmt -- --check` clean
- [x] `cargo build -p aether-hello-component --target
wasm32-unknown-unknown` builds clean
- [x] `cargo build -p aether-demo-sokoban --target
wasm32-unknown-unknown` builds clean

Phase 2 (hash-based derivation, retire `resolve_mailbox_p32`, shrink
registry) follows in a separate PR.

+++ crates/aether-component/src/raw.rs
@@ -19,13 +19,13 @@
 #[link(wasm_import_module = "aether")]
 unsafe extern "C" {
     #[link_name = "send_mail_p32"]
-    pub fn send_mail(recipient: u32, kind: u32, ptr: u32, len: u32, count: u32) -> u32;
+    pub fn send_mail(recipient: u64, kind: u32, ptr: u32, len: u32, count: u32) -> u32;
     #[link_name = "reply_mail_p32"]
     pub fn reply_mail(sender: u32, kind: u32, ptr: u32, len: u32, count: u32) -> u32;
     #[link_name = "resolve_kind_p32"]
     pub fn resolve_kind(name_ptr: u32, name_len: u32) -> u32;
     #[link_name = "resolve_mailbox_p32"]
-    pub fn resolve_mailbox(name_ptr: u32, name_len: u32) -> u32;
+    pub fn resolve_mailbox(name_ptr: u32, name_len: u32) -> u64;
     #[link_name = "save_state_p32"]
     pub fn save_state(version: u32, ptr: u32, len: u32) -> u32;
 }
@@ -34,7 +34,7 @@ unsafe extern "C" {
 /// Host-target stub for the wasm `aether::send_mail` import. Always
 /// panics — callers on non-wasm targets are misusing the SDK.
 #[cfg(not(target_arch = "wasm32"))]
-pub unsafe fn send_mail(_recipient: u32, _kind: u32, _ptr: u32, _len: u32, _count: u32) -> u32 {
+pub unsafe fn send_mail(_recipient: u64, _kind: u32, _ptr: u32, _len: u32, _count: u32) -> u32 {
     panic!("aether-component: send_mail called on non-wasm target");
 }
 
@@ -58,7 +58,7 @@ pub unsafe fn resolve_kind(_name_ptr: u32, _name_len: u32) -> u32 {
 /// Host-target stub for the wasm `aether::resolve_mailbox` import.
 /// Always panics — callers on non-wasm targets are misusing the SDK.
 #[cfg(not(target_arch = "wasm32"))]
-pub unsafe fn resolve_mailbox(_name_ptr: u32, _name_len: u32) -> u32 {
+pub unsafe fn resolve_mailbox(_name_ptr: u32, _name_len: u32) -> u64 {
     panic!("aether-component: resolve_mailbox called on non-wasm target");
 }
